ACCUMULO-94: merge from 1.3 branch
git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/trunk@1190421 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/src/main/java/org/apache/accumulo/core/client/Connector.java
@@ -140,7 +140,7 @@ public class Connector {
    * 
    * @return MultiTableBatchWriter object for configuring and writing data to
    */
-  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, int maxLatency, int maxWriteThreads) {
+  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, long maxLatency, int maxWriteThreads) {
     return impl.createMultiTableBatchWriter(maxMemory, maxLatency, maxWriteThreads);
   }
