+++ src/core/src/main/java/org/apache/accumulo/core/client/impl/ConnectorImpl.java
@@ -112,7 +112,7 @@ public class ConnectorImpl extends Connector {
   }
   
   @Override
-  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, int maxLatency, int maxWriteThreads) {
+  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, long maxLatency, int maxWriteThreads) {
     return new MultiTableBatchWriterImpl(instance, credentials, maxMemory, maxLatency, maxWriteThreads);
   }
