+++ src/core/src/main/java/org/apache/accumulo/core/client/mock/MockConnector.java
@@ -61,7 +61,7 @@ public class MockConnector extends Connector {
   }
   
   @Override
-  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, int maxLatency, int maxWriteThreads) {
+  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, long maxLatency, int maxWriteThreads) {
     return new MockMultiTableBatchWriter(acu);
   }
